Move config out of server directory
