+++ pages/_document.js
@@ -4,7 +4,7 @@ import { ServerStyleSheet } from 'styled-components';
 
 import globalStyle from 'ui/global';
 
-import { NODE_ENV, GTM_ID, GRAPH_URL } from 'server/config';
+import { NODE_ENV, GTM_ID, GRAPH_URL } from '../config';
 
 // These settings will be exposed to the world
 const clientConfig = {
